postImage + getImage
Separate the function of getting image to display from server.js (app.get()) to getImage() in controller.js and image.js router

Separate the function of posting single image to databse from createGame() to a function named postImage in controller.js. Based on property of this function, it can only be used when creating a game.